requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
behave==1.2.6
bitso-py==3.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT==0.3.2
future==0.16.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
more-itertools==7.0.0
parse==1.12.0
parse-type==0.4.2
pbr==5.1.0
pluggy==0.9.0
py==1.8.0
PyJWT==1.4.2
pytest==4.4.0
python-dateutil==2.7.4
requests==2.20.0
six==1.11.0
urllib3==1.24
websocket-client==0.40.0
Werkzeug==0.14.1
